+++ Sources/SwiftUICharts/Shared/Extensions/Extension+Hashable.swift
