+++ Sources/SwiftUICharts/Shared/Extensions/NumberFormatter+Helper.swift
