+++ Sources/SwiftUICharts/Shared/Models/ChartType.swift
@@ -1,0 +1,28 @@
+//
+//  ChartType.swift
+//  
+//
+//  Created by Will Dale on 01/01/2022.
+//
+
+import Foundation
+
+/**
+ The type of chart being used.
+ ```
+ case line // Line Chart Type
+ case bar // Bar Chart Type
+ case pie // Pie Chart Type
+ case extraLine // Extra Line Type
+ ```
+ */
+public enum ChartType {
+    /// Line Chart Type
+    case line
+    /// Bar Chart Type
+    case bar
+    /// Pie Chart Type
+    case pie
+    /// Extra Line Type
+    case extraLine
+}
