+++ Sources/SwiftUICharts/Shared/deprecated/deprecated+ColourStyle.swift
